moved cli focused methods into a separate file, and adjusted tests
