+++ clickpy/__main__.py
@@ -1,6 +1,6 @@
 """Clicker module file. This will be ran when called via 'python -m clicker'."""
 
 if __name__ == "__main__":
-    import clickpy
+    from clickpy import run
 
-    clickpy.run()
+    run()
