+++ tests/auto_click_test.py
